requirements.txt
gymnasium
stable-baselines3
pygame
numpy
matplotlib
torch
tensorboard
Pillow